[ci] sonar, y u no update pr?

+++ .ci/sonar.sh
@@ -2,6 +2,12 @@
 
 set -e
 
+# Skip sonarqube analysis if not in a pull request
+if [ "$TRAVIS_PULL_REQUEST" == "false" ]; then
+  echo "=== Not a pull request: skipping sonar analysis"
+  exit 0
+fi
+
 export SONAR_SCANNER_VERSION=4.2.0.1873
 export SONAR_SCANNER_HOME=$HOME/.sonar/sonar-scanner-$SONAR_SCANNER_VERSION-linux
 rm -rf $SONAR_SCANNER_HOME
@@ -12,17 +18,9 @@ rm $HOME/.sonar/sonar-scanner.zip
 export PATH=$SONAR_SCANNER_HOME/bin:$PATH
 export SONAR_SCANNER_OPTS="-server"
 
-# Skip sonarqube analysis if not in a pull request
-if [ "$TRAVIS_PULL_REQUEST" == "false" ]; then
-  echo "=== Not a pull request: skipping sonar analysis"
-  exit 0
-fi
 
 echo "=== Running sonar scanner"
 
-# Unset travis pre-set parameters, which are more harmful than anything
-unset SONARQUBE_SCANNER_PARAMS
-unset SONARQUBE_SKIPPED
 PACKAGE_VERSION=$(cat package.json | grep version | head -1 | awk -F: '{ print $2 }' | sed 's/[\",]//g' | tr -d '[[:space:]]')
 
 sonar-scanner \
@@ -30,6 +28,7 @@ sonar-scanner \
   -Dsonar.projectVersion="$PACKAGE_VERSION" \
   -Dsonar.organization=kuzzleio \
   -Dsonar.sources=lib \
+  -Dsonar.language=js \
   -Dsonar.cfamily.build-wrapper-output=bw-output \
   -Dsonar.host.url=https://sonarcloud.io \
   -Dsonar.login="$SONARCLOUD_TOKEN" \
@@ -37,11 +36,5 @@ sonar-scanner \
   -Dsonar.pullrequest.branch="$TRAVIS_PULL_REQUEST_BRANCH" \
   -Dsonar.pullrequest.base="$TRAVIS_BRANCH" \
   -Dsonar.pullrequest.provider="GitHub" \
+  -Dsonar.pullrequest.github.repository="$TRAVIS_REPO_SLUG" \
   -X
-
-  # -Dsonar.login="$SONARCLOUD_TOKEN" \
-  # -Dsonar.language=js \
-  # -Dsonar.pullrequest.github.repository="$TRAVIS_REPO_SLUG" \
-  # -Dsonar.github.oauth="$SONAR_GITHUB_TOKEN" \
-  # -Dsonar.github.repository="$TRAVIS_REPO_SLUG" \
-  # -X
